+++ OutOfSchool/OutOfSchool.WebApi/Models/AverageRatingDTO.cs
@@ -8,7 +8,7 @@ public class AverageRatingDto
 {
     public float Rate { get; set; }
 
-    public long RateQuantity { get; set; }
+    public int RateQuantity { get; set; }
 
     public Guid EntityId { get; set; }
 }
